requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-decouple==3.8
python-dotenv==1.0.0
python-steam-api==1.0.8
requests==2.31.0
soupsieve==2.4.1
steamwebapi==0.1.4
telebot==0.0.5
urllib3==2.0.3
